requirements.txt
-e .
pytest
aws-cdk.core
